+++ frontend/pages/devices/_deviceId/parameters/_parameterId/edit.vue
@@ -59,12 +59,12 @@ permissions and limitations under the Licence.
       </v-card-actions>
     </v-card>
     <v-subheader
-      v-if="deviceParameters.length > 1"
+      v-if="deviceParametersSortedAlphabetically.length > 1"
     >
       Existing parameters
     </v-subheader>
     <BaseList
-      :list-items="deviceParameters"
+      :list-items="deviceParametersSortedAlphabetically"
     >
       <template #list-item="{item,index}">
         <ParameterListItem
@@ -79,7 +79,7 @@ permissions and limitations under the Licence.
 
 <script lang="ts">
 import { Component, mixins } from 'nuxt-property-decorator'
-import { mapActions, mapState } from 'vuex'
+import { mapActions, mapGetters, mapState } from 'vuex'
 
 import CheckEditAccess from '@/mixins/CheckEditAccess'
 
@@ -109,7 +109,8 @@ import SaveAndCancelButtons from '@/components/shared/SaveAndCancelButtons.vue'
   },
   computed: {
     ...mapState('vocabulary', ['units']),
-    ...mapState('devices', ['deviceParameter', 'deviceParameters'])
+    ...mapState('devices', ['deviceParameter']),
+    ...mapGetters('devices', ['deviceParametersSortedAlphabetically'])
   },
   methods: {
     ...mapActions('devices', ['updateDeviceParameter', 'loadDeviceParameters', 'loadDeviceParameter']),
@@ -122,7 +123,7 @@ export default class ParametersEditPage extends mixins(CheckEditAccess) {
 
   // vuex definition for typescript check
   deviceParameter!: DevicesState['deviceParameter']
-  deviceParameters!: DevicesState['deviceParameters']
+  deviceParametersSortedAlphabetically!: DevicesState['deviceParameters']
   loadDeviceParameter!: LoadDeviceParameterAction
   loadDeviceParameters!: LoadDeviceParametersAction
   updateDeviceParameter!: UpdateDeviceParameterAction