+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Added:
 - Tooltip to explain why a paramter can not be deleted if the delete button is disabled
 - Shortcut to edit or delete a value of a parameter of device, platform or configuration
 
+Changed:
+- Parameters of device, platform or configuration are now sorted alphabetically
+
 Fixed:
 - Adjusted the ordering of the elements in sensorML